generator/requirements.txt
Faker==4.0.3
python-dateutil==2.8.1
six==1.14.0
text-unidecode==1.3
fake-useragent==0.1.11